Change `changelog` task to for direct invocation from external command
This primarily changes the function signature of the `changelog` task to
always take the revision to start from and the name of the next
tag/version.

In order to be called externally however, an ember-cli project instance
needs to be passed in, which conflicted with the existing 'project'
config object. This meant two additional changes:

1. The arguments of helpers to avoid using `this` to make it clear that
the project instance may not have originated from the `changelog`
command.
2. The existing `project` is now merged with the user config to make
one master config object with all relevant options.